make server_client interface
