+++ internal/telegram/bot.go
@@ -3,22 +3,19 @@ package telegram
 import (
 	"fmt"
 	"log"
-
-	sc "github.com/valyankilyan/sandbox-market-bot/internal/telegram/server_client"
 )
 
 type Bot interface {
-	New(token string)
 }
 
 type TBot struct {
 	token  string
-	server *sc.Server
+	server Server
 }
 
 var apiAddr string = "https://api.telegram.org/bot%s/%s"
 
-func New(token string, server *sc.Server) *TBot {
+func New(token string, server Server) *TBot {
 	if token == "" {
 		log.Fatal("No telegram token was given.")
 	}
@@ -34,4 +31,9 @@ func (b *TBot) requestURL(command string) (string, error) {
 		return "", fmt.Errorf("no command was given in requestURL")
 	}
 	return fmt.Sprintf(apiAddr, b.token, command), nil
+}
+
+type Server interface {
+	CreateUser(User)
+	UpdateTinkoffToken(user User, token string) error
 }
