+++ internal/telegram/datastruct.go
@@ -1,11 +1,16 @@
 package telegram
 
-import "github.com/valyankilyan/sandbox-market-bot/internal/telegram/server_client"
-
+type User struct {
+	TgId         int64
+	IsBot        bool
+	FirstName    string
+	Username     string
+	LanguageCode string
+}
 type message struct {
 	MessageID int64
 	Chat      chat
-	From      server_client.User
+	From      User
 	Date      int64
 	Text      string
 }
