+++ internal/telegram/server_client/user.go
@@ -3,21 +3,30 @@ package server_client
 import (
 	"log"
 
+	"github.com/valyankilyan/sandbox-market-bot/internal/telegram"
 	srv "github.com/valyankilyan/sandbox-market-bot/pkg/server_api"
 )
 
-func (s *Server) CreateUser(usr User) {
+func (s *Server) CreateUser(usr telegram.User) {
 	nwusr := srv.User{
 		Name:       usr.FirstName,
 		TgUserName: usr.Username,
-		TgId:       usr.ID,
+		TgId:       usr.TgId,
 	}
 
 	s.client.CreateUser(s.ctx, &nwusr)
 }
 
-func (s *Server) getUserByTg(user User) *srv.User {
-	usr, err := s.client.ReadUser(s.ctx, &srv.TgId{TgId: user.ID})
+func (s *Server) UpdateTinkoffToken(user telegram.User, token string) error {
+	usr := s.getUserByTg(user)
+	usr.TinkoffToken = token
+	_, err := s.client.UpdateUser(s.ctx, usr)
+	log.Println("couldn't update User tinkoff token:", err)
+	return err
+}
+
+func (s *Server) getUserByTg(user telegram.User) *srv.User {
+	usr, err := s.client.ReadUser(s.ctx, &srv.TgId{TgId: user.TgId})
 	if err != nil {
 		log.Println("error in updateTinkofToken:", err)
 		return &srv.User{}
@@ -31,16 +40,8 @@ func (s *Server) getUserByTg(user User) *srv.User {
 	return usr
 }
 
-func (s *Server) UpdateTinkoffToken(user User, token string) error {
-	usr := s.getUserByTg(user)
-	usr.TinkoffToken = token
-	_, err := s.client.UpdateUser(s.ctx, usr)
-	log.Println("couldn't update User tinkoff token:", err)
-	return err
-}
-
-func (s *Server) ReadUserToken(user User) (token string, err error) {
-	usr, err := s.client.ReadUser(s.ctx, &srv.TgId{TgId: user.ID})
+func (s *Server) readUserToken(user telegram.User) (token string, err error) {
+	usr, err := s.client.ReadUser(s.ctx, &srv.TgId{TgId: user.TgId})
 	if err != nil {
 		log.Println("user don't have tinkoff token:", err)
 		return "", err
